+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Move most account settings into the Account adapter
 * Allow the user to set a custom success message to be displayed after they verify an AnVIL account
+* Move the setting to specify the template for account link verification emails into the AccountAdapter
 
 ## 0.25.0 (2024-08-07)